[WIP] Minor code cleanup in packages, renamed classes, added skeleton of Spock test.
