+++ src/main/java/com/evocalize/dnslookup/Application.java
@@ -4,10 +4,10 @@ import org.springframework.boot.SpringApplication;
 import org.springframework.boot.autoconfigure.SpringBootApplication;
 
 @SpringBootApplication
-public class DnslookupApplication {
+public class Application {
 
     public static void main(String[] args) {
-        SpringApplication.run(DnslookupApplication.class, args);
+        SpringApplication.run(Application.class, args);
     }
 
 }
